Fix stub files
