--- moveit_py/moveit/__init__.pyi
@@ -1,3 +1,0 @@
-from moveit_py.core import *
-from moveit_py.planning import *
-from moveit_py.utils import get_launch_params_filepath as get_launch_params_filepath
