+++ moveit_py/moveit/core/__init__.pyi
@@ -1,0 +1,8 @@
+from typing import Any
+
+class JointModelGroup:
+    def __init__(self, *args, **kwargs) -> None: ...
+    @property
+    def joint_model_names(self) -> Any: ...
+    @property
+    def name(self) -> Any: ...
