+++ moveit_py/moveit/core/kinematic_constraints.pyi
@@ -1,0 +1,5 @@
+from typing import Any
+
+def construct_constraints_from_node(*args, **kwargs) -> Any: ...
+def construct_joint_constraint(*args, **kwargs) -> Any: ...
+def construct_link_constraint(*args, **kwargs) -> Any: ...
