+++ moveit_py/moveit/core/planning_interface.pyi
@@ -1,0 +1,15 @@
+from typing import Any
+
+class MotionPlanResponse:
+    def __init__(self, *args, **kwargs) -> None: ...
+    def __bool__(self) -> Any: ...
+    @property
+    def error_code(self) -> Any: ...
+    @property
+    def planner_id(self) -> Any: ...
+    @property
+    def planning_time(self) -> Any: ...
+    @property
+    def start_state(self) -> Any: ...
+    @property
+    def trajectory(self) -> Any: ...
