+++ moveit_py/moveit/core/robot_model.pyi
@@ -1,0 +1,17 @@
+from typing import Any
+
+class RobotModel:
+    def __init__(self, *args, **kwargs) -> None: ...
+    def get_joint_model_group(self, *args, **kwargs) -> Any: ...
+    def get_model_info(self, *args, **kwargs) -> Any: ...
+    def has_joint_model_group(self, *args, **kwargs) -> Any: ...
+    @property
+    def joint_model_group_names(self) -> Any: ...
+    @property
+    def joint_model_groups(self) -> Any: ...
+    @property
+    def model_frame(self) -> Any: ...
+    @property
+    def name(self) -> Any: ...
+    @property
+    def root_joint_name(self) -> Any: ...
