+++ moveit_py/moveit/planning.pyi
@@ -13,10 +13,14 @@ class LockedPlanningSceneContextManagerRW:
 class MoveItPy:
     def __init__(self, *args, **kwargs) -> None: ...
     def get_planning_component(self, *args, **kwargs) -> Any: ...
-    def get_planning_pipeline_names(self, *args, **kwargs) -> Any: ...
     def get_planning_scene_monitor(self, *args, **kwargs) -> Any: ...
     def get_robot_model(self, *args, **kwargs) -> Any: ...
     def shutdown(self, *args, **kwargs) -> Any: ...
+
+class MultiPipelinePlanRequestParameters:
+    def __init__(self, *args, **kwargs) -> None: ...
+    @property
+    def multi_plan_request_parameters(self) -> Any: ...
 
 class PlanRequestParameters:
     max_acceleration_scaling_factor: Any
@@ -27,22 +31,12 @@ class PlanRequestParameters:
     planning_time: Any
     def __init__(self, *args, **kwargs) -> None: ...
 
-class PlanSolution:
-    def __init__(self, *args, **kwargs) -> None: ...
-    def __bool__(self) -> Any: ...
-    @property
-    def error_code(self) -> Any: ...
-    @property
-    def start_state(self) -> Any: ...
-    @property
-    def trajectory(self) -> Any: ...
-
 class PlanningComponent:
     def __init__(self, *args, **kwargs) -> None: ...
     def execute(self, *args, **kwargs) -> Any: ...
     def get_start_state(self, *args, **kwargs) -> Any: ...
     def plan(self, *args, **kwargs) -> Any: ...
-    def set_goal(self, *args, **kwargs) -> Any: ...
+    def set_goal_state(self, *args, **kwargs) -> Any: ...
     def set_path_constraints(self, *args, **kwargs) -> Any: ...
     def set_start_state(self, *args, **kwargs) -> Any: ...
     def set_start_state_to_current_state(self, *args, **kwargs) -> Any: ...
